refect: 점수 실수 입력
- 정수 입력 범위에서 실수 입력 범위로 전환
- DTO 필드이름 grade로 변경
- 매퍼 타겟이름 변경

+++ src/main/java/com/cockatoo/domain/grade/api/GradeController.java
@@ -34,7 +34,7 @@ public class GradeController {
     }
 
     @GetMapping("/evaluate")
-    ResponseEntity<EvaluateGradeResponse> evaluateGrade(@PathParam("score") Long score) {
+    ResponseEntity<EvaluateGradeResponse> evaluateGrade(@PathParam("score") Double score) {
         EvaluateGradeResponse response = gradeService.evaluateGrade(score);
         return ResponseEntity.status(HttpStatus.OK).body(response);
     }
